+++ rolo/testing/pytest.py
@@ -13,10 +13,10 @@ from werkzeug import serving
 
 from rolo import Router
 from rolo.asgi import ASGIAdapter, ASGILifespanListener
-from rolo.dispatcher import handler_dispatcher
 from rolo.gateway import Gateway
 from rolo.gateway.asgi import AsgiGateway
 from rolo.gateway.wsgi import WsgiGateway
+from rolo.routing import handler_dispatcher
 from rolo.serving.twisted import TwistedGateway
 from rolo.websocket.adapter import WebSocketListener
