+++ tests/test_resource.py
@@ -2,7 +2,7 @@ import pytest
 from werkzeug.exceptions import MethodNotAllowed
 
 from rolo import Request, Resource, Response, Router, resource
-from rolo.dispatcher import handler_dispatcher
+from rolo.routing import handler_dispatcher
 
 
 class TestResource:
